Make the mirror data mapping much better.
Now it replaces common charactere mismatches (microsoft double-quotes for example), and ignores common things placed in the CSV name field that would not match the real name.

Additionally, we have a name remapping syustem so that Armor +1 matches +1 Armor properly.

Also, added style to the nodes on item display screen so people do not miss it.